refactor: local .npmrc
