+++ .changeset/chilly-garlics-joke.md
@@ -1,0 +1,5 @@
+---
+"@pnpm/config": minor
+---
+
+New function added: `readLocalConfig(dir: string)`.
