+++ packages/config/package.json
@@ -32,6 +32,7 @@
   },
   "homepage": "https://github.com/pnpm/pnpm/blob/main/packages/config#readme",
   "dependencies": {
+    "@npmcli/config": "6.1.0",
     "@pnpm/constants": "workspace:*",
     "@pnpm/error": "workspace:*",
     "@pnpm/git-utils": "workspace:*",
@@ -41,6 +42,7 @@
     "@pnpm/read-project-manifest": "workspace:*",
     "@pnpm/types": "workspace:*",
     "camelcase": "^6.3.0",
+    "camelcase-keys": "^6.2.2",
     "can-write-to-dir": "^1.1.1",
     "is-subdir": "^1.2.0",
     "is-windows": "^1.0.2",
@@ -48,6 +50,7 @@
     "path-absolute": "^1.0.1",
     "path-name": "^1.0.0",
     "ramda": "npm:@pnpm/ramda@0.28.1",
+    "read-ini-file": "^3.1.0",
     "realpath-missing": "^1.1.0",
     "which": "^2.0.2"
   },
