+++ packages/plugin-commands-rebuild/src/recursive.ts
@@ -1,20 +1,18 @@
-import path from 'path'
 import {
   RecursiveSummary,
   throwOnCommandFail,
 } from '@pnpm/cli-utils'
 import {
   Config,
+  readLocalConfig,
 } from '@pnpm/config'
 import { arrayOfWorkspacePackagesToMap } from '@pnpm/find-workspace-packages'
 import { logger } from '@pnpm/logger'
 import { sortPackages } from '@pnpm/sort-packages'
 import { createOrConnectStoreController, CreateStoreControllerOptions } from '@pnpm/store-connection-manager'
 import { Project, ProjectManifest } from '@pnpm/types'
-import camelcaseKeys from 'camelcase-keys'
 import mem from 'mem'
 import pLimit from 'p-limit'
-import readIniFile from 'read-ini-file'
 import { rebuildProjects as rebuildAll, RebuildOptions, rebuildSelectedPkgs } from './implementation'
 
 type RecursiveRebuildOpts = CreateStoreControllerOptions & Pick<Config,
@@ -162,47 +160,4 @@ export async function recursiveRebuild (
   }
 
   throwOnFail(result)
-}
-
-async function readLocalConfig (prefix: string) {
-  try {
-    const ini = await readIniFile(path.join(prefix, '.npmrc')) as Record<string, string>
-    const config = camelcaseKeys(ini) as (Record<string, string> & { hoist?: boolean })
-    if (config.shamefullyFlatten) {
-      config.hoistPattern = '*'
-      // TODO: print a warning
-    }
-    if (config.hoist === false) {
-      config.hoistPattern = ''
-    }
-
-    for (const [key, val] of Object.entries(config)) {
-      if (typeof val !== 'string') {
-        continue
-      }
-
-      config[key] = parseEnv(val, process.env)
-    }
-
-    return config
-  } catch (err: any) { // eslint-disable-line
-    if (err.code !== 'ENOENT') throw err
-    return {}
-  }
-}
-
-// https://github.com/npm/cli/blob/706b3d3f227de43a095263926d2eef2b4e4cf2a9/workspaces/config/lib/env-replace.js
-function parseEnv (field: string, env: NodeJS.ProcessEnv) {
-  const envExpr = /(?<!\\)(\\*)\$\{([^${}]+)\}/g
-
-  return field.replace(envExpr, (orig, esc: string, name: string) => {
-    const val = env[name] !== undefined ? env[name] : `$\{${name}}`
-
-    // consume the escape chars that are relevant.
-    if (esc.length % 2) {
-      return orig.slice((esc.length + 1) / 2)
-    }
-
-    return `${(esc.slice(esc.length / 2))}${val}` // eslint-disable-line @typescript-eslint/restrict-template-expressions
-  })
 }
